Create the header of the website
The main edits occured in the css file.
h1, banner-text, and button text sizes are adjusted.
Image size for logo and banner image is adjusted to
appropriate sizes. Header itself is turned into a flex
display with appropriate padding. Proper formatting
was done to the list elements for the links in the top of
the header so that they are not underlined and are white.