Add quote section to the website
Change font size and font according to the template given.
Also decided not to use flexbox for this particular part
of the website since I wanted the quote credentials part
to be on the right side. There might be a way to do this with
flexbox as well, could be researched in the future.